Fix flawed benchmark
Creation: RegExp are lazily setup internally, thus the benchmark must
force a setup by calling the exec() method, so as to enable
meaningful comparison with other collections.

Look-up: there is apparently some optimizations in RegExp.exec() which
causes the method to return much faster when the argument used is one
which was used recently, as if the RegExp implementation keeps a mini
internal cache of argument-result (just speculating). When feeding the
RegExp() with a random argument, the benchmarking results are different.
I revised the lookup benchmarks to use only random argument.